+++ quanlynhahang/src/main/java/poly/edu/quanlynhahang/dto/PublicProductResponse.java
@@ -38,7 +38,7 @@ public record PublicProductResponse(
         }
     }
 
-    private static BigDecimal money(Double value) {
-        return value == null ? null : BigDecimal.valueOf(value).setScale(2, RoundingMode.HALF_UP);
+    private static BigDecimal money(BigDecimal value) {
+        return value == null ? null : value.setScale(2, RoundingMode.HALF_UP);
     }
 }
